+++ test/report.test.ts
@@ -1,6 +1,7 @@
 // (C) 2007-2019 GoodData Corporation
 import "isomorphic-fetch";
 import fetchMock from "fetch-mock";
+import { AFM, ExecuteAFM } from "@gooddata/typings";
 import { ReportModule } from "../src/report";
 import { XhrModule, ApiResponseError } from "../src/xhr";
 import { IExportConfig, IExportResponse } from "../src/interfaces";
@@ -27,6 +28,125 @@ describe("report", () => {
         });
 
         describe("exportResult", () => {
+            it("should sanitized showFilters config", () => {
+                fetchMock.mock(projectUri, {
+                    status: SUCCESS_REQUEST_STATUS,
+                    body: { uri: createdReport },
+                });
+
+                const mockTask = (status: number) => ({ status, uri: createdReport });
+                const finishedTask = mockTask(SUCCESS_REQUEST_STATUS);
+                const runningTask = mockTask(ACCEPTED_REQUEST_STATUS);
+                mockPollingRequest(createdReport, runningTask, finishedTask);
+
+                const showFilters: AFM.CompatibilityFilter[] = [
+                    {
+                        positiveAttributeFilter: {
+                            displayForm: {
+                                uri: "bar",
+                            },
+                            in: ["/gdc/md/bar1", "/gdc/md/bar2"],
+                        },
+                    },
+                    {
+                        negativeAttributeFilter: {
+                            displayForm: {
+                                identifier: "foo",
+                            },
+                            notIn: ["foo1", "foo2"],
+                            textFilter: true,
+                        },
+                    },
+                    {
+                        absoluteDateFilter: {
+                            dataSet: {
+                                uri: "/gdc/md/i6k6sk4sznefv1kf0f2ls7jf8tm5ida6/obj/330",
+                            },
+                            from: "2011-01-01",
+                            to: "2011-12-31",
+                        },
+                    },
+                    {
+                        relativeDateFilter: {
+                            to: 0,
+                            from: -3,
+                            granularity: "GDC.time.quarter",
+                            dataSet: {
+                                uri: "/gdc/md/myproject/obj/921",
+                            },
+                        },
+                    },
+                ];
+
+                const expectedShowFilters: ExecuteAFM.CompatibilityFilter[] = [
+                    {
+                        positiveAttributeFilter: {
+                            displayForm: {
+                                uri: "bar",
+                            },
+                            in: {
+                                uris: ["/gdc/md/bar1", "/gdc/md/bar2"],
+                            },
+                        },
+                    },
+                    {
+                        negativeAttributeFilter: {
+                            displayForm: {
+                                identifier: "foo",
+                            },
+                            notIn: {
+                                values: ["foo1", "foo2"],
+                            },
+                        },
+                    },
+                    {
+                        absoluteDateFilter: {
+                            dataSet: {
+                                uri: "/gdc/md/i6k6sk4sznefv1kf0f2ls7jf8tm5ida6/obj/330",
+                            },
+                            from: "2011-01-01",
+                            to: "2011-12-31",
+                        },
+                    },
+                    {
+                        relativeDateFilter: {
+                            dataSet: {
+                                uri: "/gdc/md/myproject/obj/921",
+                            },
+                            from: -3,
+                            granularity: "GDC.time.quarter",
+                            to: 0,
+                        },
+                    },
+                ];
+
+                const exportConfig: IExportConfig = {
+                    title: "title",
+                    format: "xlsx",
+                    mergeHeaders: false,
+                    showFilters,
+                };
+
+                return mockedReportModule()
+                    .exportResult(projectId, executionResult, exportConfig, { pollStep: 1 })
+                    .then(() => {
+                        const [, settings] = fetchMock.lastCall(
+                            `/gdc/internal/projects/${projectId}/exportResult`,
+                        );
+                        expect(JSON.parse(settings.body as string)).toEqual({
+                            resultExport: {
+                                executionResult: "/executionResult/1234",
+                                exportConfig: {
+                                    title: "title",
+                                    format: "xlsx",
+                                    mergeHeaders: false,
+                                    showFilters: expectedShowFilters,
+                                },
+                            },
+                        });
+                    });
+            });
+
             it("should return created file", () => {
                 fetchMock.mock(projectUri, {
                     status: SUCCESS_REQUEST_STATUS,